refactor: html 페이지 반환하는 로직 한 곳에 모으기
