+++ src/main/java/shopping/controller/ViewController.java
@@ -1,18 +1,19 @@
 package shopping.controller;
 
-import java.util.List;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import shopping.application.ProductService;
 import shopping.dto.response.ProductResponse;
 
+import java.util.List;
+
 @Controller
-public class ProductController {
+public class ViewController {
 
     private final ProductService productService;
 
-    public ProductController(ProductService productService) {
+    public ViewController(ProductService productService) {
         this.productService = productService;
     }
 
@@ -21,5 +22,15 @@ public class ProductController {
         List<ProductResponse> products = productService.findAll();
         model.addAttribute("products", products);
         return "index";
+    }
+
+    @GetMapping("/login")
+    public String loginPage() {
+        return "login";
+    }
+
+    @GetMapping("/cart")
+    public String getCartPage() {
+        return "cart";
     }
 }
